Final Hackathon/agent_backend/requirements.txt
fastapi 
httpx 
beautifulsoup4 
langchain 
langchain-google-genai 
python-dotenv
langchain-community
tavily-python
uvicorn
langchain-tavily
motor